ollama-docker/code/requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-text-splitters
langchain-core
langchain-classic
chromadb